+++ app/src/main/java/com/dicoding/picodiploma/sejiwaproject/MainActivity.kt
@@ -4,6 +4,8 @@ import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
+import com.dicoding.picodiploma.sejiwaproject.adapter.LeagueAdapter
+import com.dicoding.picodiploma.sejiwaproject.db.LeagueData
 import com.dicoding.picodiploma.sejiwaproject.model.League
 import org.jetbrains.anko.*
 import org.jetbrains.anko.recyclerview.v7.recyclerView
