+++ app/src/main/java/com/dicoding/picodiploma/sejiwaproject/adapter/LeagueAdapter.kt
@@ -1,4 +1,4 @@
-package com.dicoding.picodiploma.sejiwaproject
+package com.dicoding.picodiploma.sejiwaproject.adapter
 
 import android.content.Intent
 import android.view.View
@@ -8,6 +8,7 @@ import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import com.bumptech.glide.request.RequestOptions
+import com.dicoding.picodiploma.sejiwaproject.R
 import com.dicoding.picodiploma.sejiwaproject.detail.DetailLeagueActivity
 import com.dicoding.picodiploma.sejiwaproject.model.League
 import com.dicoding.picodiploma.sejiwaproject.view.ItemLeagueUI
